requirements.txt
# Core API libraries
openai>=1.0.0
python-dotenv>=1.0.0

# HTTP client and HTML parsing
aiohttp>=3.8.0
beautifulsoup4>=4.12.0

# Browser automation for price scraping
playwright>=1.40.0

# Data processing and CSV handling
pandas>=2.0.0
python-dateutil>=2.8.0

# Sports data for individual player game logs
sportsdataverse>=0.0.39

# XGBoost pinned to avoid binary format incompatibility with sportsdataverse
xgboost<3.0,>=1.7.0

# HTML scraping no longer needed - using ESPN API

# Text similarity for semantic deduplication
fuzzywuzzy>=0.18.0
python-levenshtein>=0.12.0

# Email delivery via SendGrid HTTP API (more reliable than SMTP)
sendgrid>=6.11.0

# Tunnel Fits Multi-Player Expansion Dependencies
oxylabs>=1.0.0  # Oxylabs SDK for Google Lens visual search (replaces SerpApi)
Pillow>=10.0.0  # Image processing for quality scoring
imagehash>=4.3.0  # Perceptual image hashing for deduplication

# Development (optional)
pytest>=7.0.0
black>=23.0.0